refactor(racingcar.domain): 입출력 기능 병합으로 패키지 구조 변경 및 클래스명 변경
