--- src/main/java/racingcar/message/ErrorMessage.java
@@ -1,5 +1,0 @@
-package racingcar.message;
-
-
-public enum ErrorMessage {
-}
