--- src/main/java/racingcar/message/GameMessage.java
@@ -1,5 +1,0 @@
-package racingcar.message;
-
-
-public enum GameMessage {
-}
